qt-cpp: Move test out of src/
